+++ CHANGELOG.md
@@ -2,6 +2,7 @@ v0.2.0
 ======
 
 * Support multiple metrics updated in a single request (newline-separated)
+* Gauge metric type (`g`)
 
 v0.1.0: Sep 14 2011
 ===================